refactor(managers): 去掉门面模块的冗余 _manager 后缀
app/managers/ 包名已表达「管理器」语义，文件名再带 _manager 后缀属 stutter
（app.managers.notification_manager 三处 manager 重复）。

- git mv 去后缀：notification/downloader/mediaserver/mediarecognize/storage.py（R100 纯重命名）
- 类名保留 NotificationManager 等（Notification/Storage 太泛、易与 schema 类型混淆）
- 调用处改走 __init__ 再导出：chain 与 5 个门面测试套统一 from app.managers import X，
  文件名不再出现在调用处（同 from app.core.module import ModuleManager 范式）
- _LAZY_EXPORTS/TYPE_CHECKING 与 modules docstring 文案同步去后缀

全量 1738 passed / 19 skipped
