+++ app/modules/qbittorrent/__init__.py
@@ -42,7 +42,7 @@ class QbittorrentModule(_ModuleBase, _DownloaderBase[Qbittorrent]):
     """
     Qbittorrent 下载器模块（app.modules.IDownloader 后端）。
 
-    作为 Downloader 域后端，由门面 app.managers.downloader_manager.DownloaderManager 统一分发调用。
+    作为 Downloader 域后端，由门面 app.managers.downloader.DownloaderManager 统一分发调用。
 
     .. deprecated::
         经 ChainBase.run_module("download"/"list_torrents"/…) 的字符串 ABI 分发为 v2 兼容路径，
