+++ tests/test_storage_facade.py
@@ -7,7 +7,7 @@
 from unittest import TestCase
 from unittest.mock import MagicMock, patch
 
-from app.managers.storage_manager import StorageManager
+from app.managers import StorageManager
 from app.schemas.exception import RateLimitExceededException
